requirements.txt
streamlit
torch>=2.0.0
timm>=0.9.0
Pillow
torchvision
